Add references to tables and figure (#271)
As per the feedback received on avenue, changes were made to the SRS tex
file to ensure that tables and figrues were properly referenced with
captions and working links. Preambles to the related sections and
subsections were also added to allow such referencing since some
subsections only contained a diagram/table. The revision table was
updated to reflect these changes.

closes #271